Make TestAction conform ModelConvertible
