+++ Sources/TuistLoader/ModelConvertible/Settings+ModelConvertible.swift
@@ -5,10 +5,10 @@ import TuistCore
 
 extension TuistCore.Settings: ModelConvertible {
     typealias BuildConfigurationTuple = (TuistCore.BuildConfiguration, TuistCore.Configuration?)
-    
+
     init(manifest: ProjectDescription.Settings, generatorPaths: GeneratorPaths) throws {
         let base = try manifest.base.mapValues { try TuistCore.SettingValue(manifest: $0, generatorPaths: generatorPaths) }
-        
+
         let configurations = try manifest.configurations
             .reduce([TuistCore.BuildConfiguration: TuistCore.Configuration?]()) { acc, val in
                 var result = acc
@@ -17,13 +17,13 @@ extension TuistCore.Settings: ModelConvertible {
                     result[variant] = try TuistCore.Configuration(manifest: configuration, generatorPaths: generatorPaths)
                 }
                 return result
-        }
+            }
         let defaultSettings = try TuistCore.DefaultSettings(manifest: manifest.defaultSettings, generatorPaths: generatorPaths)
         self.init(base: base,
                   configurations: configurations,
                   defaultSettings: defaultSettings)
     }
-    
+
     fileprivate static func buildConfigurationTuple(from customConfiguration: CustomConfiguration,
                                                     generatorPaths: GeneratorPaths) throws -> BuildConfigurationTuple {
         let buildConfiguration = try TuistCore.BuildConfiguration(manifest: customConfiguration, generatorPaths: generatorPaths)
